i#1684 xarch-IR: Remove mangle.c from drdecode
Moves the handful of functions from mangle.c used by drdecode (things
that are difficult to untangle from the IR functions) into ir_utils.c
Renames the top-level ir_utils.c to ir_utils_shared.c.  Adds
ir_utils.h.

Issue: #1684
