+++ core/arch/aarchxx/mangle.c
@@ -1,5 +1,5 @@
 /* **********************************************************
- * Copyright (c) 2014-2019 Google, Inc.  All rights reserved.
+ * Copyright (c) 2014-2020 Google, Inc.  All rights reserved.
  * Copyright (c) 2016 ARM Limited. All rights reserved.
  * **********************************************************/
 
@@ -91,127 +91,6 @@ mangle_arch_init(void)
 #endif
 }
 
-byte *
-remangle_short_rewrite(dcontext_t *dcontext, instr_t *instr, byte *pc, app_pc target)
-{
-#ifdef AARCH64
-    ASSERT_NOT_IMPLEMENTED(false); /* FIXME i#1569 */
-    return NULL;
-#else
-    uint mangled_sz = CTI_SHORT_REWRITE_LENGTH;
-    uint raw_jmp;
-    ASSERT(instr_is_cti_short_rewrite(instr, pc));
-    if (target == NULL)
-        target = decode_raw_jmp_target(dcontext, pc + CTI_SHORT_REWRITE_B_OFFS);
-    instr_set_target(instr, opnd_create_pc(target));
-    instr_allocate_raw_bits(dcontext, instr, mangled_sz);
-    instr_set_raw_bytes(instr, pc, mangled_sz);
-    encode_raw_jmp(dr_get_isa_mode(dcontext), target, (byte *)&raw_jmp,
-                   pc + CTI_SHORT_REWRITE_B_OFFS);
-    instr_set_raw_word(instr, CTI_SHORT_REWRITE_B_OFFS, raw_jmp);
-    instr_set_operands_valid(instr, true);
-    return (pc + mangled_sz);
-#endif
-}
-
-instr_t *
-convert_to_near_rel_arch(dcontext_t *dcontext, instrlist_t *ilist, instr_t *instr)
-{
-#ifdef AARCH64
-    ASSERT_NOT_IMPLEMENTED(false); /* FIXME i#1569 */
-    return NULL;
-#else
-    int opcode = instr_get_opcode(instr);
-    if (opcode == OP_b_short) {
-        instr_set_opcode(instr, OP_b);
-        return instr;
-    } else if (opcode == OP_cbz || opcode == OP_cbnz) {
-        /* While for non-trace-mode we could get by w/o converting,
-         * as we use local stubs with a far-away link-through-stub
-         * soln needed even for regular branches and thus these would
-         * reach the stub, they won't reach for traces.
-         * Thus we mirror what x86 does for jecxz:
-         *       cbz foo
-         *  =>
-         *       cbnz fall
-         *       jmp foo
-         *  fall:
-         *
-         * The fact that we invert the cbr ends up requiring extra logic
-         * in linkstub_cbr_disambiguate().
-         */
-        app_pc target = NULL;
-        uint mangled_sz, offs, raw_jmp;
-        reg_id_t src_reg;
-
-        if (ilist != NULL) {
-            /* PR 266292: for meta instrs, insert separate instrs */
-            opnd_t tgt = instr_get_target(instr);
-            instr_t *fall = INSTR_CREATE_label(dcontext);
-            instr_t *jmp = INSTR_CREATE_b(dcontext, tgt);
-            ASSERT(instr_is_meta(instr));
-            /* reverse order */
-            instrlist_meta_postinsert(ilist, instr, fall);
-            instrlist_meta_postinsert(ilist, instr, jmp);
-            instrlist_meta_postinsert(ilist, instr, instr);
-            instr_set_target(instr, opnd_create_instr(fall));
-            instr_invert_cbr(instr);
-            return jmp; /* API specifies we return the long-reach cti */
-        }
-
-        if (opnd_is_near_pc(instr_get_target(instr)))
-            target = opnd_get_pc(instr_get_target(instr));
-        else if (opnd_is_near_instr(instr_get_target(instr))) {
-            instr_t *tgt = opnd_get_instr(instr_get_target(instr));
-            /* XXX: not using get_app_instr_xl8() b/c drdecodelib doesn't link
-             * mangle_shared.c.
-             */
-            target = instr_get_translation(tgt);
-            if (target == NULL && instr_raw_bits_valid(tgt))
-                target = instr_get_raw_bits(tgt);
-            ASSERT(target != NULL);
-        } else
-            ASSERT_NOT_REACHED();
-
-        /* PR 251646: cti_short_rewrite: target is in src0, so operands are
-         * valid, but raw bits must also be valid, since they hide the multiple
-         * instrs.  For x64, it is marked for re-relativization, but it's
-         * special since the target must be obtained from src0 and not
-         * from the raw bits (since that might not reach).
-         */
-        /* query IR before we set raw bits */
-        ASSERT(opnd_is_reg(instr_get_src(instr, 1)));
-        src_reg = opnd_get_reg(instr_get_src(instr, 1));
-        /* need 6 bytes */
-        mangled_sz = CTI_SHORT_REWRITE_LENGTH;
-        instr_allocate_raw_bits(dcontext, instr, mangled_sz);
-        offs = 0;
-        /* first 2 bytes: cbz or cbnz to "cur pc" + 2 which means immed is 1 */
-        instr_set_raw_byte(instr, offs, 0x08 | (src_reg - DR_REG_R0));
-        offs++;
-        instr_set_raw_byte(instr, offs, (opcode == OP_cbz) ? CBNZ_BYTE_A : CBZ_BYTE_A);
-        offs++;
-        /* next 4 bytes: b to target */
-        ASSERT(offs == CTI_SHORT_REWRITE_B_OFFS);
-        encode_raw_jmp(dr_get_isa_mode(dcontext),
-                       instr->bytes + offs /*not target, b/c may not reach*/,
-                       (byte *)&raw_jmp, instr->bytes + offs);
-        instr_set_raw_word(instr, offs, raw_jmp);
-        offs += sizeof(int);
-        ASSERT(offs == mangled_sz);
-        LOG(THREAD, LOG_INTERP, 2, "convert_to_near_rel: cbz/cbnz opcode\n");
-        /* original target operand is still valid */
-        instr_set_operands_valid(instr, true);
-        return instr;
-    }
-    ASSERT_NOT_REACHED();
-    return instr;
-#endif
-}
-
-/***************************************************************************/
-#if !defined(STANDALONE_DECODER)
-
 void
 insert_clear_eflags(dcontext_t *dcontext, clean_call_info_t *cci, instrlist_t *ilist,
                     instr_t *instr)
@@ -221,9 +100,9 @@ insert_clear_eflags(dcontext_t *dcontext, clean_call_info_t *cci, instrlist_t *i
      */
 }
 
-#    ifdef AARCH64
+#ifdef AARCH64
 /* Maximum positive immediate offset for STP/LDP with 64 bit registers. */
-#        define MAX_STP_OFFSET 504
+#    define MAX_STP_OFFSET 504
 
 /* Creates a memory reference for registers saved/restored to memory. */
 static opnd_t
@@ -383,7 +262,7 @@ insert_restore_inline_registers(dcontext_t *dcontext, instrlist_t *ilist, instr_
                                      (callee_info_t *)ci);
 }
 
-#    endif
+#endif
 
 /* Pushes not only the GPRs but also simd regs, xip, and xflags, in
  * priv_mcontext_t order.
@@ -404,9 +283,9 @@ insert_push_all_registers(dcontext_t *dcontext, clean_call_info_t *cci,
                               _IF_AARCH64(bool out_of_line))
 {
     uint dstack_offs = 0;
-#    ifdef AARCH64
+#ifdef AARCH64
     uint max_offs;
-#    endif
+#endif
     if (cci == NULL)
         cci = &default_clean_call_info;
     ASSERT(proc_num_simd_registers() == MCXT_NUM_SIMD_SLOTS);
@@ -416,7 +295,7 @@ insert_push_all_registers(dcontext_t *dcontext, clean_call_info_t *cci,
          */
     }
     /* FIXME i#1551: once we have cci->num_simd_skip, skip this if possible */
-#    ifdef AARCH64
+#ifdef AARCH64
     /* X0 is used to hold the stack pointer. */
     cci->reg_skip[DR_REG_X0 - DR_REG_START_GPR] = false;
     /* X1 and X2 are used to save and restore the status and control registers. */
@@ -539,7 +418,7 @@ insert_push_all_registers(dcontext_t *dcontext, clean_call_info_t *cci,
         INSTR_CREATE_ldr(dcontext, opnd_create_reg(DR_REG_X2),
                          opnd_create_base_disp(DR_REG_SP, DR_REG_NULL, 0,
                                                REG_OFFSET(DR_REG_X2), OPSZ_8)));
-#    else
+#else
 
     /* vstmdb always does writeback */
     PRE(ilist, instr,
@@ -620,7 +499,7 @@ insert_push_all_registers(dcontext_t *dcontext, clean_call_info_t *cci,
     dstack_offs += XSP_SZ;
     ASSERT(cci->skip_save_flags || cci->num_simd_skip != 0 || cci->num_regs_skip != 0 ||
            dstack_offs == (uint)get_clean_call_switch_stack_size());
-#    endif
+#endif
     return dstack_offs;
 }
 
@@ -634,7 +513,7 @@ insert_pop_all_registers(dcontext_t *dcontext, clean_call_info_t *cci, instrlist
 {
     if (cci == NULL)
         cci = &default_clean_call_info;
-#    ifdef AARCH64
+#ifdef AARCH64
     uint current_offs;
     /* mov x0, sp */
     PRE(ilist, instr,
@@ -712,7 +591,7 @@ insert_pop_all_registers(dcontext_t *dcontext, clean_call_info_t *cci, instrlist
                              OPND_CREATE_INT16(get_clean_call_switch_stack_size())));
     }
 
-#    else
+#else
     /* We rely on dr_set_mcontext_priv() to set the app's stolen reg value,
      * and the stack swap to set the sp value: we assume the stolen reg on
      * the stack still has our TLS base in it.
@@ -754,16 +633,16 @@ insert_pop_all_registers(dcontext_t *dcontext, clean_call_info_t *cci, instrlist
     PRE(ilist, instr,
         INSTR_CREATE_vldm_wb(dcontext, OPND_CREATE_MEMLIST(DR_REG_SP), SIMD_REG_LIST_LEN,
                              SIMD_REG_LIST_16_31));
-#    endif
+#endif
 }
 
-#    ifndef AARCH64
+#ifndef AARCH64
 reg_id_t
 shrink_reg_for_param(reg_id_t regular, opnd_t arg)
 {
     return regular;
 }
-#    endif /* !AARCH64 */
+#endif /* !AARCH64 */
 
 /* Return true if opnd is a register, but not XSP, or immediate zero on AArch64. */
 static bool
@@ -807,7 +686,7 @@ insert_parameter_preparation(dcontext_t *dcontext, instrlist_t *ilist, instr_t *
         uint n = num_args - NUM_REGPARM;
         /* On both ARM and AArch64 the stack pointer is kept (2 * XSP_SZ)-aligned. */
         stack_inc = ALIGN_FORWARD(n, 2) * XSP_SZ;
-#    ifdef AARCH64
+#ifdef AARCH64
         for (i = 0; i < n; i += 2) {
             opnd_t *arg0 = &args[NUM_REGPARM + i];
             opnd_t *arg1 = i + 1 < n ? &args[NUM_REGPARM + i + 1] : NULL;
@@ -871,7 +750,7 @@ insert_parameter_preparation(dcontext_t *dcontext, instrlist_t *ilist, instr_t *
                         opnd_is_reg(*arg1) ? *arg1 : opnd_create_reg(DR_REG_XZR)));
             }
         }
-#    else /* ARM */
+#else /* ARM */
         /* XXX: We could use OP_stm here, but with lots of awkward corner cases. */
         PRE(ilist, instr,
             INSTR_CREATE_sub(dcontext, opnd_create_reg(DR_REG_XSP),
@@ -887,7 +766,7 @@ insert_parameter_preparation(dcontext_t *dcontext, instrlist_t *ilist, instr_t *
                                        arg));
             }
         }
-#    endif
+#endif
     }
 
     /* Initialise regs[], which encodes the contents of parameter registers.
@@ -1033,7 +912,7 @@ int
 insert_out_of_line_context_switch(dcontext_t *dcontext, instrlist_t *ilist,
                                   instr_t *instr, bool save, byte *encode_pc)
 {
-#    ifdef AARCH64
+#ifdef AARCH64
     if (save) {
         /* Reserve stack space to push the context. We do it here instead of
          * in insert_push_all_registers, so we can save the original value
@@ -1081,10 +960,10 @@ insert_out_of_line_context_switch(dcontext_t *dcontext, instrlist_t *ilist,
     }
 
     return get_clean_call_switch_stack_size();
-#    else
+#else
     ASSERT_NOT_IMPLEMENTED(false); /* FIXME i#1621: NYI on AArch32. */
     return 0;
-#    endif
+#endif
 }
 
 /*###########################################################################
@@ -1098,7 +977,7 @@ static void
 mangle_stolen_reg(dcontext_t *dcontext, instrlist_t *ilist, instr_t *instr,
                   instr_t *next_instr, bool instr_to_be_removed);
 
-#    ifndef AARCH64
+#ifndef AARCH64
 
 /* i#1662 optimization: we try to pick the same scratch register during
  * mangling to provide more opportunities for optimization,
@@ -1132,17 +1011,17 @@ find_prior_scratch_reg_restore(dcontext_t *dcontext, instr_t *instr, reg_id_t *p
     return NULL;
 }
 
-#    endif /* !AARCH64 */
+#endif /* !AARCH64 */
 
 /* optimized spill: only if not immediately spilled already */
 static void
 insert_save_to_tls_if_necessary(dcontext_t *dcontext, instrlist_t *ilist, instr_t *where,
                                 reg_id_t reg, ushort slot)
 {
-#    ifdef AARCH64
+#ifdef AARCH64
     /* FIXME i#1569: not yet optimized */
     PRE(ilist, where, instr_create_save_to_tls(dcontext, reg, slot));
-#    else
+#else
     instr_t *prev;
     reg_id_t prior_reg;
     DEBUG_DECLARE(bool tls;)
@@ -1162,10 +1041,10 @@ insert_save_to_tls_if_necessary(dcontext_t *dcontext, instrlist_t *ilist, instr_
     } else {
         PRE(ilist, where, instr_create_save_to_tls(dcontext, reg, slot));
     }
-#    endif
+#endif
 }
 
-#    ifndef AARCH64
+#ifndef AARCH64
 
 /* If instr is inside an IT block, removes it from the block and
  * leaves it as an isolated (un-encodable) predicated instr, with any
@@ -1321,95 +1200,7 @@ mangle_reinstate_it_blocks(dcontext_t *dcontext, instrlist_t *ilist, instr_t *st
     });
 }
 
-#    endif /* !AARCH64 */
-
-#endif /* !STANDALONE_DECODER */
-/* We export these mov/push utilities to drdecode */
-
-void
-insert_mov_immed_arch(dcontext_t *dcontext, instr_t *src_inst, byte *encode_estimate,
-                      ptr_int_t val, opnd_t dst, instrlist_t *ilist, instr_t *instr,
-                      OUT instr_t **first, OUT instr_t **last)
-{
-#ifdef AARCH64
-    instr_t *mov;
-    int i;
-
-    CLIENT_ASSERT(opnd_is_reg(dst), "AArch64 cannot store an immediate direct to memory");
-
-    if (opnd_get_reg(dst) == DR_REG_XZR) {
-        /* Moving a value to the zero register is a no-op. We insert nothing,
-         * so *first and *last are set to NULL. Caller beware!
-         */
-        if (first != NULL)
-            *first = NULL;
-        if (last != NULL)
-            *last = NULL;
-        return;
-    }
-
-    ASSERT((uint)(opnd_get_reg(dst) - DR_REG_X0) < 31);
-    if (src_inst != NULL)
-        val = (ptr_int_t)encode_estimate;
-
-    /* movz x(dst), #(val & 0xffff) */
-    mov = INSTR_CREATE_movz(dcontext, dst, OPND_CREATE_INT16(val & 0xffff),
-                            OPND_CREATE_INT8(0));
-    PRE(ilist, instr, mov);
-    if (first != NULL)
-        *first = mov;
-    for (i = 1; i < 4; i++) {
-        if ((val >> (16 * i) & 0xffff) != 0) {
-            /* movk x(dst), #(val >> sh & 0xffff), lsl #(sh) */
-            mov = INSTR_CREATE_movk(dcontext, dst,
-                                    OPND_CREATE_INT16((val >> 16 * i) & 0xffff),
-                                    OPND_CREATE_INT8(i * 16));
-            PRE(ilist, instr, mov);
-        }
-    }
-    if (last != NULL)
-        *last = mov;
-#else
-    instr_t *mov1, *mov2;
-    if (src_inst != NULL)
-        val = (ptr_int_t)encode_estimate;
-    CLIENT_ASSERT(opnd_is_reg(dst), "ARM cannot store an immediate direct to memory");
-    /* MVN writes the bitwise inverse of an immediate value to the dst register */
-    /* XXX: we could check for larger tile/rotate immed patterns */
-    if (src_inst == NULL && ~val >= 0 && ~val <= 0xff) {
-        mov1 = INSTR_CREATE_mvn(dcontext, dst, OPND_CREATE_INT(~val));
-        PRE(ilist, instr, mov1);
-        mov2 = NULL;
-    } else {
-        /* To use INT16 here and pass the size checks in opnd_create_immed_int
-         * we'd have to add UINT16 (or sign-extend the bottom half again):
-         * simpler to use INT, and our general ARM philosophy is to use INT and
-         * ignore immed sizes at instr creation time (only at encode time do we
-         * check them).
-         */
-        mov1 = INSTR_CREATE_movw(dcontext, dst,
-                                 (src_inst == NULL)
-                                     ? OPND_CREATE_INT(val & 0xffff)
-                                     : opnd_create_instr_ex(src_inst, OPSZ_2, 0));
-        PRE(ilist, instr, mov1);
-        val = (val >> 16) & 0xffff;
-        if (val == 0) {
-            /* movw zero-extends so we're done */
-            mov2 = NULL;
-        } else {
-            mov2 = INSTR_CREATE_movt(dcontext, dst,
-                                     (src_inst == NULL)
-                                         ? OPND_CREATE_INT(val)
-                                         : opnd_create_instr_ex(src_inst, OPSZ_2, 16));
-            PRE(ilist, instr, mov2);
-        }
-    }
-    if (first != NULL)
-        *first = mov1;
-    if (last != NULL)
-        *last = mov2;
-#endif
-}
+#endif /* !AARCH64 */
 
 void
 patch_mov_immed_arch(dcontext_t *dcontext, ptr_int_t val, byte *pc, instr_t *first,
@@ -1417,16 +1208,6 @@ patch_mov_immed_arch(dcontext_t *dcontext, ptr_int_t val, byte *pc, instr_t *fir
 {
     ASSERT_NOT_IMPLEMENTED(false); /* FIXME i#1551, i#1569 */
 }
-
-void
-insert_push_immed_arch(dcontext_t *dcontext, instr_t *src_inst, byte *encode_estimate,
-                       ptr_int_t val, instrlist_t *ilist, instr_t *instr,
-                       OUT instr_t **first, OUT instr_t **last)
-{
-    ASSERT_NOT_IMPLEMENTED(false); /* FIXME i#1551, i#1569 */
-}
-
-#ifndef STANDALONE_DECODER /* back for rest of file */
 
 /* Used for fault translation */
 bool
@@ -1463,7 +1244,7 @@ mangle_syscall_arch(dcontext_t *dcontext, instrlist_t *ilist, uint flags, instr_
     ASSERT(DR_REG_STOLEN_MIN > DR_REG_SYSNUM);
 }
 
-#    ifdef UNIX
+#ifdef UNIX
 /* Inserts code to handle clone into ilist.
  * instr is the syscall instr itself.
  * Assumes that instructions exist beyond instr in ilist.
@@ -1490,7 +1271,7 @@ mangle_insert_clone_code(dcontext_t *dcontext, instrlist_t *ilist, instr_t *inst
     instr_set_meta(instr_get_prev(in));
     PRE(ilist, in, parent);
 }
-#    endif /* UNIX */
+#endif /* UNIX */
 
 void
 mangle_interrupt(dcontext_t *dcontext, instrlist_t *ilist, instr_t *instr,
@@ -1499,7 +1280,7 @@ mangle_interrupt(dcontext_t *dcontext, instrlist_t *ilist, instr_t *instr,
     ASSERT_NOT_IMPLEMENTED(false); /* FIXME i#1551, i#1569 */
 }
 
-#    ifndef AARCH64
+#ifndef AARCH64
 
 /* Adds a mov of the fall-through address into IBL_TARGET_REG, predicated
  * with the inverse of instr's predicate.
@@ -1555,13 +1336,13 @@ app_instr_is_in_it_block(dcontext_t *dcontext, instr_t *instr)
     return (instr_get_isa_mode(instr) == DR_ISA_ARM_THUMB && instr_is_predicated(instr));
 }
 
-#    endif /* !AARCH64 */
+#endif /* !AARCH64 */
 
 instr_t *
 mangle_direct_call(dcontext_t *dcontext, instrlist_t *ilist, instr_t *instr,
                    instr_t *next_instr, bool mangle_calls, uint flags)
 {
-#    ifdef AARCH64
+#ifdef AARCH64
     ptr_int_t target, retaddr;
 
     ASSERT(instr_get_opcode(instr) == OP_bl);
@@ -1573,7 +1354,7 @@ mangle_direct_call(dcontext_t *dcontext, instrlist_t *ilist, instr_t *instr,
     instrlist_remove(ilist, instr); /* remove OP_bl */
     instr_destroy(dcontext, instr);
     return next_instr;
-#    else
+#else
     /* Strategy: replace OP_bl with 2-step mov immed into lr + OP_b */
     ptr_uint_t retaddr;
     uint opc = instr_get_opcode(instr);
@@ -1636,14 +1417,14 @@ mangle_direct_call(dcontext_t *dcontext, instrlist_t *ilist, instr_t *instr,
     if (in_it)
         mangle_reinstate_it_blocks(dcontext, ilist, bound_start, next_instr);
     return next_instr;
-#    endif
+#endif
 }
 
 instr_t *
 mangle_indirect_call(dcontext_t *dcontext, instrlist_t *ilist, instr_t *instr,
                      instr_t *next_instr, bool mangle_calls, uint flags)
 {
-#    ifdef AARCH64
+#ifdef AARCH64
     ASSERT(instr_get_opcode(instr) == OP_blr);
     PRE(ilist, instr,
         instr_create_save_to_tls(dcontext, IBL_TARGET_REG, IBL_TARGET_SLOT));
@@ -1662,7 +1443,7 @@ mangle_indirect_call(dcontext_t *dcontext, instrlist_t *ilist, instr_t *instr,
     instrlist_remove(ilist, instr); /* remove OP_blr */
     instr_destroy(dcontext, instr);
     return next_instr;
-#    else
+#else
     ptr_uint_t retaddr;
     bool in_it = app_instr_is_in_it_block(dcontext, instr);
     instr_t *bound_start = INSTR_CREATE_label(dcontext);
@@ -1703,7 +1484,7 @@ mangle_indirect_call(dcontext_t *dcontext, instrlist_t *ilist, instr_t *instr,
     if (in_it)
         mangle_reinstate_it_blocks(dcontext, ilist, bound_start, next_instr);
     return next_instr;
-#    endif
+#endif
 }
 
 void
@@ -1718,7 +1499,7 @@ instr_t *
 mangle_indirect_jump(dcontext_t *dcontext, instrlist_t *ilist, instr_t *instr,
                      instr_t *next_instr, uint flags)
 {
-#    ifdef AARCH64
+#ifdef AARCH64
     ASSERT(instr_get_opcode(instr) == OP_br || instr_get_opcode(instr) == OP_ret);
     PRE(ilist, instr,
         instr_create_save_to_tls(dcontext, IBL_TARGET_REG, IBL_TARGET_SLOT));
@@ -1735,7 +1516,7 @@ mangle_indirect_jump(dcontext_t *dcontext, instrlist_t *ilist, instr_t *instr,
     instrlist_remove(ilist, instr); /* remove OP_br or OP_ret */
     instr_destroy(dcontext, instr);
     return next_instr;
-#    else
+#else
     bool remove_instr = false;
     int opc = instr_get_opcode(instr);
     dr_isa_mode_t isa_mode = instr_get_isa_mode(instr);
@@ -1765,7 +1546,7 @@ mangle_indirect_jump(dcontext_t *dcontext, instrlist_t *ilist, instr_t *instr,
         opnd_set_size(&memop, OPSZ_VAR_REGLIST);
         instr_set_src(instr, 0, memop);
         instr_set_dst(instr, 0, opnd_create_reg(IBL_TARGET_REG));
-#        ifdef CLIENT_INTERFACE
+#    ifdef CLIENT_INTERFACE
         /* We target only the typical return instructions: multi-pop here */
         if (TEST(INSTR_CLOBBER_RETADDR, instr->flags) && opc == OP_ldmia) {
             bool writeback = instr_num_srcs(instr) > 1;
@@ -1780,7 +1561,7 @@ mangle_indirect_jump(dcontext_t *dcontext, instrlist_t *ilist, instr_t *instr,
                      XINST_CREATE_store(dcontext, memop, opnd_create_reg(dr_reg_stolen)));
             } /* else not a pop */
         }
-#        endif
+#    endif
     } else if (opc == OP_bx || opc == OP_bxj) {
         ASSERT(opnd_is_reg(instr_get_target(instr)));
         if (opnd_same(instr_get_target(instr), opnd_create_reg(dr_reg_stolen))) {
@@ -1902,7 +1683,7 @@ mangle_indirect_jump(dcontext_t *dcontext, instrlist_t *ilist, instr_t *instr,
              */
             mangle_stolen_reg(dcontext, ilist, instr, immed_next, remove_instr);
         }
-#        ifdef CLIENT_INTERFACE
+#    ifdef CLIENT_INTERFACE
         /* We target only the typical return instructions: single pop here */
         if (TEST(INSTR_CLOBBER_RETADDR, instr->flags) && opc == OP_ldr) {
             bool writeback = instr_num_srcs(instr) > 1;
@@ -1914,7 +1695,7 @@ mangle_indirect_jump(dcontext_t *dcontext, instrlist_t *ilist, instr_t *instr,
                      XINST_CREATE_store(dcontext, memop, opnd_create_reg(dr_reg_stolen)));
             } /* else not a pop */
         }
-#        endif
+#    endif
     }
     if (instr_is_predicated(instr)) {
         mangle_add_predicated_fall_through(dcontext, ilist, instr, next_instr,
@@ -1928,7 +1709,7 @@ mangle_indirect_jump(dcontext_t *dcontext, instrlist_t *ilist, instr_t *instr,
     if (in_it)
         mangle_reinstate_it_blocks(dcontext, ilist, bound_start, next_instr);
     return next_instr;
-#    endif
+#endif
 }
 
 /* Local single-instr-window scratch reg picker.  Only considers r0-r3, so the
@@ -1950,7 +1731,7 @@ pick_scratch_reg(dcontext_t *dcontext, instr_t *instr, bool dead_reg_ok,
     if (should_restore != NULL)
         *should_restore = true;
 
-#    ifndef AARCH64 /* FIXME i#1569: not yet optimized */
+#ifndef AARCH64 /* FIXME i#1569: not yet optimized */
     if (find_prior_scratch_reg_restore(dcontext, instr, &reg) != NULL &&
         reg != REG_NULL && !instr_uses_reg(instr, reg) &&
         /* Ensure no conflict in scratch regs for PC or stolen reg
@@ -1965,7 +1746,7 @@ pick_scratch_reg(dcontext_t *dcontext, instr_t *instr, bool dead_reg_ok,
             LOG(THREAD, LOG_INTERP, 4, "use last scratch reg %s\n", reg_names[reg]);
         });
     } else
-#    endif
+#endif
         reg = REG_NULL;
 
     if (reg == REG_NULL) {
@@ -2011,7 +1792,7 @@ instr_t *
 mangle_rel_addr(dcontext_t *dcontext, instrlist_t *ilist, instr_t *instr,
                 instr_t *next_instr)
 {
-#    ifdef AARCH64
+#ifdef AARCH64
     uint opc = instr_get_opcode(instr);
     opnd_t dst = instr_get_dst(instr, 0);
     opnd_t src = instr_get_src(instr, 0);
@@ -2063,7 +1844,7 @@ mangle_rel_addr(dcontext_t *dcontext, instrlist_t *ilist, instr_t *instr,
     instrlist_remove(ilist, instr);
     instr_destroy(dcontext, instr);
     return NULL;
-#    else
+#else
     /* Compute the value of r15==pc for orig app instr */
     ptr_int_t r15 =
         (ptr_int_t)decode_cur_pc(instr_get_raw_bits(instr), instr_get_isa_mode(instr),
@@ -2130,10 +1911,10 @@ mangle_rel_addr(dcontext_t *dcontext, instrlist_t *ilist, instr_t *instr,
         mangle_reinstate_it_blocks(dcontext, ilist, bound_start, next_instr);
     }
     return next_instr;
-#    endif
+#endif
 }
 
-#    ifndef AARCH64
+#ifndef AARCH64
 
 /* mangle simple pc read, pc read in gpr_list is handled in mangle_gpr_list_read */
 static void
@@ -2169,7 +1950,7 @@ mangle_pc_read(dcontext_t *dcontext, instrlist_t *ilist, instr_t *instr,
         PRE(ilist, next_instr, instr_create_restore_from_tls(dcontext, reg, slot));
 }
 
-#    endif /* !AARCH64 */
+#endif /* !AARCH64 */
 
 /* save tls_base from dr_reg_stolen to reg and load app value to dr_reg_stolen */
 static void
@@ -2248,7 +2029,7 @@ mangle_stolen_reg(dcontext_t *dcontext, instrlist_t *ilist, instr_t *instr,
      */
     ASSERT(!instr_is_meta(instr) && instr_uses_reg(instr, dr_reg_stolen));
 
-#    ifndef AARCH64 /* FIXME i#1569: recognise "move" on AArch64 */
+#ifndef AARCH64 /* FIXME i#1569: recognise "move" on AArch64 */
     /* optimization, convert simple mov to ldr/str:
      * - "mov r0  -> r10"  ==> "str r0 -> [r10_slot]"
      * - "mov r10 -> r0"   ==> "ldr [r10_slot] -> r0"
@@ -2280,7 +2061,7 @@ mangle_stolen_reg(dcontext_t *dcontext, instrlist_t *ilist, instr_t *instr,
             ASSERT_NOT_REACHED();
         }
     }
-#    endif
+#endif
 
     /* move stolen reg value into tmp reg for app instr execution */
     tmp = pick_scratch_reg(dcontext, instr, false, &slot, &should_restore);
@@ -2301,7 +2082,7 @@ instr_t *
 mangle_reads_thread_register(dcontext_t *dcontext, instrlist_t *ilist, instr_t *instr,
                              instr_t *next_instr)
 {
-#    ifdef AARCH64
+#ifdef AARCH64
     reg_id_t reg = opnd_get_reg(instr_get_dst(instr, 0));
     ASSERT(instr->opcode == OP_mrs);
     if (reg != dr_reg_stolen) {
@@ -2321,7 +2102,7 @@ mangle_reads_thread_register(dcontext_t *dcontext, instrlist_t *ilist, instr_t *
     instrlist_remove(ilist, instr);
     instr_destroy(dcontext, instr);
     return next_instr;
-#    else
+#else
     opnd_t opnd;
     reg_id_t reg;
     bool in_it = app_instr_is_in_it_block(dcontext, instr);
@@ -2364,10 +2145,10 @@ mangle_reads_thread_register(dcontext_t *dcontext, instrlist_t *ilist, instr_t *
     if (in_it)
         mangle_reinstate_it_blocks(dcontext, ilist, bound_start, next_instr);
     return next_instr;
-#    endif
+#endif
 }
 
-#    ifdef AARCH64
+#ifdef AARCH64
 instr_t *
 mangle_writes_thread_register(dcontext_t *dcontext, instrlist_t *ilist, instr_t *instr,
                               instr_t *next_instr)
@@ -2392,9 +2173,9 @@ mangle_writes_thread_register(dcontext_t *dcontext, instrlist_t *ilist, instr_t 
     instr_destroy(dcontext, instr);
     return next_instr;
 }
-#    endif
+#endif
 
-#    ifndef AARCH64
+#ifndef AARCH64
 
 static void
 store_reg_to_memlist(dcontext_t *dcontext, instrlist_t *ilist, instr_t *instr,
@@ -2912,9 +2693,9 @@ mangle_gpr_list_write(dcontext_t *dcontext, instrlist_t *ilist, instr_t *instr,
     return next_instr;
 }
 
-#    endif /* !AARCH64 */
+#endif /* !AARCH64 */
 
-#    ifdef AARCH64
+#ifdef AARCH64
 /* We mangle a conditional branch that uses the stolen register like this:
  *
  *     cbz   x28, target     # x28 is stolen register
@@ -2972,7 +2753,7 @@ mangle_cbr_stolen_reg(dcontext_t *dcontext, instrlist_t *ilist, instr_t *instr,
     PRE(ilist, next_instr, fall);
     PRE(ilist, next_instr, instr_create_restore_from_tls(dcontext, reg, slot));
 }
-#    endif /* AARCH64 */
+#endif /* AARCH64 */
 
 /* On ARM, we need mangle app instr accessing registers pc and dr_reg_stolen.
  * We use this centralized mangling routine here to handle complex issues with
@@ -2982,7 +2763,7 @@ instr_t *
 mangle_special_registers(dcontext_t *dcontext, instrlist_t *ilist, instr_t *instr,
                          instr_t *next_instr)
 {
-#    ifdef AARCH64
+#ifdef AARCH64
     if (!instr_uses_reg(instr, dr_reg_stolen))
         return next_instr;
     if (instr_is_cbr(instr))
@@ -2990,7 +2771,7 @@ mangle_special_registers(dcontext_t *dcontext, instrlist_t *ilist, instr_t *inst
     else if (!instr_is_mbr(instr))
         mangle_stolen_reg(dcontext, ilist, instr, instr_get_next(instr), false);
     return next_instr;
-#    else
+#else
     bool finished = false;
     bool in_it =
         instr_get_isa_mode(instr) == DR_ISA_ARM_THUMB && instr_is_predicated(instr);
@@ -3036,7 +2817,7 @@ mangle_special_registers(dcontext_t *dcontext, instrlist_t *ilist, instr_t *inst
         mangle_reinstate_it_blocks(dcontext, ilist, bound_start, bound_end);
     }
     return next_instr;
-#    endif
+#endif
 }
 
 void
@@ -3046,7 +2827,7 @@ float_pc_update(dcontext_t *dcontext)
     ASSERT_NOT_REACHED();
 }
 
-#    ifdef AARCH64
+#ifdef AARCH64
 instr_t *
 mangle_icache_op(dcontext_t *dcontext, instrlist_t *ilist, instr_t *instr,
                  instr_t *next_instr, app_pc pc)
@@ -3123,12 +2904,9 @@ mangle_icache_op(dcontext_t *dcontext, instrlist_t *ilist, instr_t *instr,
         ASSERT_NOT_REACHED();
     return next_instr;
 }
-#    endif
+#endif
 
 /* END OF CONTROL-FLOW MANGLING ROUTINES
  *###########################################################################
  *###########################################################################
  */
-
-#endif /* !STANDALONE_DECODER */
-/***************************************************************************/
